exam.cls reference
For those who are looking to compile the tex files from scratch, it is obvious they'll need to find out how to install plugins on their own in order to get things to work...however, if they are looking for the `exam.cls` file that will provide the styling for them, they'll need the right version in order for things to compile well and include all right control sequences. I propose we make the link for that available here in the resources.